+++ include/RE/NiMath.h
@@ -4,7 +4,7 @@
 namespace RE
 {
 	constexpr float NI_INFINITY = FLT_MAX;
-	constexpr float NI_PI = 3.1415926535897932;
+	constexpr float NI_PI = static_cast<float>(3.1415926535897932);
 	constexpr float NI_HALF_PI = 0.5 * NI_PI;
 	constexpr float NI_TWO_PI = 2.0 * NI_PI;
