+++ src/RE/TESAIForm.cpp
@@ -29,7 +29,7 @@ namespace RE
 
 	UInt8 TESAIForm::GetEnergyLevel() const
 	{
-		return pun_bits(aiData.energyLevel1, aiData.energyLevel2, aiData.energyLevel3, aiData.energyLevel4, aiData.energyLevel5, aiData.energyLevel6, aiData.energyLevel7, aiData.energyLevel8);
+		return static_cast<UInt8>(pun_bits(aiData.energyLevel1, aiData.energyLevel2, aiData.energyLevel3, aiData.energyLevel4, aiData.energyLevel5, aiData.energyLevel6, aiData.energyLevel7, aiData.energyLevel8));
 	}
